# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/system/activity/user/{userId}": {
      "get": {
        "operationId": "SystemController_getActivitiesByUser",
        "summary": "Get activities by user ID",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User activities retrieved"
          }
        },
        "tags": [
          "System"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "patient",
                "administrative_staff",
                "superuser",
                "ambulance",
                "doctor",
                "nurse"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User creation result (check success field)"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/stats/overview": {
      "get": {
        "operationId": "UsersController_getStats",
        "summary": "Get user statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User statistics retrieved successfully"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/role/{role}": {
      "get": {
        "operationId": "UsersController_findByRole",
        "summary": "Get users by role",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/search/{query}": {
      "get": {
        "operationId": "UsersController_search",
        "summary": "Search users",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findById",
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved (check success field for not-found)"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User update result (check success field)"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_patchUpdate",
        "summary": "Partial update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User update result (check success field)"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deletion result (check success field)"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus",
        "summary": "Update user status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status update result (check success field)"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NexCare Hospital Management API",
    "description": "Complete REST API for the NexCare Hospital Administrative Operations Platform. Features role-based access control (RBAC), JWT authentication, and comprehensive modules for patients, appointments, billing, beds, ambulance, inventory, feedback, and system management.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "x-user-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "User role for RBAC (superuser, administrative_staff, patient, ambulance, doctor, nurse)"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Smith",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "jane@nexcare.com",
            "description": "Email address"
          },
          "role": {
            "type": "string",
            "enum": [
              "patient",
              "administrative_staff",
              "superuser",
              "ambulance",
              "doctor",
              "nurse"
            ],
            "example": "doctor",
            "description": "User role"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "patientId": {
            "type": "string",
            "example": "P1002",
            "description": "Associated patient ID if applicable"
          },
          "dept": {
            "type": "string",
            "example": "Cardiology",
            "description": "Department (for doctors/staff)"
          }
        },
        "required": [
          "name",
          "email",
          "role",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Smith",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "jane@nexcare.com",
            "description": "Email address"
          },
          "role": {
            "type": "string",
            "enum": [
              "patient",
              "administrative_staff",
              "superuser",
              "ambulance",
              "doctor",
              "nurse"
            ],
            "example": "doctor",
            "description": "User role"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "On Leave"
            ],
            "example": "Active",
            "description": "User status"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "patientId": {
            "type": "string",
            "example": "P1002",
            "description": "Associated patient ID if applicable"
          },
          "dept": {
            "type": "string",
            "example": "Cardiology",
            "description": "Department (for doctors/staff)"
          }
        }
      },
...
    }
  }
}